Refactored controllers into their own package
